requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
appdirs==1.4.4
astroid==3.1.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs==23.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cityscapesScripts==2.2.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm @ file:///croot/comm_1671231121260/work
contourpy==1.1.1
crcmod==1.7
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.8
docker-pycreds==0.4.0
dsdl==0.1.10.11
environs==9.5.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2024.2.0
ftfy @ file:///tmp/build/80754af9/ftfy_1598643610002/work
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.21.4
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib-resources==6.1.1
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipython @ file:///croot/ipython_1691532092695/work
ipywidgets @ file:///croot/ipywidgets_1679394798311/work
isort==5.13.2
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2==3.1.3
jmespath==0.10.0
joblib @ file:///croot/joblib_1685113087166/work
jsonmodels==2.6.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1700168618520/work
kiwisolver==1.4.5
ldm==0.1.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.7.4
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mccabe==0.7.0
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv==2.1.0
mmengine==0.10.3
model-index==0.1.11
mpmath==1.3.0
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==3.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.0.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///croot/pillow_1696580024257/work
pkgutil_resolve_name==1.3.10
platformdirs @ file:///croot/platformdirs_1692205439124/work
pooch @ file:///croot/pooch_1695850093751/work
prettytable==3.9.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
Pygments @ file:///croot/pygments_1684279966437/work
pylint==3.1.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
referencing==0.34.0
regex @ file:///croot/regex_1696515298636/work
requests==2.28.2
rich==13.4.2
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn @ file:///croot/scikit-learn_1694788527225/work
scipy==1.10.1
sentry-sdk==2.0.1
setproctitle==1.3.3
shapely==2.0.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2023.7.10
timm==0.9.16
tomli==2.0.1
tomlkit==0.12.4
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.65.2
traitlets @ file:///croot/traitlets_1671143879854/work
triton==2.1.0
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3 @ file:///croot/urllib3_1698257533958/work
wandb==0.16.6
wcwidth==0.2.13
widgetsnbextension @ file:///croot/widgetsnbextension_1679313860248/work
yapf==0.40.2
zipp @ file:///croot/zipp_1704206909481/work
